pyproject.toml
[tool.poetry]
name = "cv-builder-api"
version = "0.1.0"
description = ""
authors = ["minsau <minsau@prototipo-devs.com.mx>"]

[tool.poetry.dependencies]
python = "==3.7.6"
asgiref= "==3.3.4"
Django= "==3.2.2"
django-filter= "==2.4.0"
djangorestframework= "==3.12.4"
djangorestframework-simplejwt= "==4.6.0"
PyJWT= "==2.1.0"
pytz= "==2021.1"
sqlparse= "==0.4.1"
typing-extensions= "==3.10.0.0"
poetry = "^1.1.6"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "*"
pipenv-to-requirements = "*"
pre-commit = "*"
pylint = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"